Remove SQl injection using paramaterized query, Prepared Statements and NamedParameterJdbcTemplate codes are written using libraries of Spring framework
