+++ Wissen/src/com/wissen/named/parameter/jdbc/template/AppConfig.java
@@ -1,0 +1,24 @@
+package com.wissen.named.parameter.jdbc.template;
+
+// Manually create and configure Jdbc Template:
+
+/*
+ * import org.springframework.jdbc.core.JdbcTemplate;
+ * import org.springframework.jdbc.datasource.DriverManagerDataSource;
+ * import javax.sql.DataSource;
+ * 
+ * @Configuration
+ * public class CreateManuallyConfigureJdbcTemplate {
+ *  
+ *  	@Bean
+ *  	public DataSource dataSource() {
+ *  		DriverManagerDataSource dataSouce = new DriverManagerDataSource();
+ *  		dataSouce.setDriverClassName("com.mysql.cj.jdbc.Driver");
+ *  		dataSource.setUrl("jdbc:mysql://localhost:3306/your_database");
+ *  		dataSource.setUsername("admin");
+	        dataSource.setPassword("admin");
+	        return dataSource;
+ *  	}
+ *  	    
+ * }
+ */
