+++ Wissen/src/com/wissen/named/parameter/jdbc/template/JdbcTemplateConfig.java
@@ -1,0 +1,12 @@
+package com.wissen.named.parameter.jdbc.template;
+
+/*
+ * @Configuration 
+ * public class JdbcTemplateConfig {
+ * 	
+ * 		@Bean
+ * 		public NamedParameterJdbcTemplate namedParameterJdbcTemplate(DataSource dataSource) {
+ * 			return new NamedParameterJdbcTemplate(dataSource);
+ * 		}
+ * }
+ */
